+++ NEWS.md
@@ -1,5 +1,7 @@
 # covr (development version)
 
+* Generated files from [cpp11](https://cpp11.r-lib.org/) are now ignored (#437, @jimhester)
+
 * `codecov()` and `coveralls()` now retry failed requests before raising an error (#428, @jameslamb)
 
 # covr 3.5.0